+++ .pnp.js
@@ -393,6 +393,19 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "linkType": "HARD",
         }]
       ]],
+      ["@arkecosystem/crypto-identities", [
+        ["npm:1.1.0", {
+          "packageLocation": "./.yarn/cache/@arkecosystem-crypto-identities-npm-1.1.0-7b4f463ad0-2.zip/node_modules/@arkecosystem/crypto-identities/",
+          "packageDependencies": [
+            ["@arkecosystem/crypto-identities", "npm:1.1.0"],
+            ["bcrypto", "npm:5.3.0"],
+            ["bstring", "npm:0.3.9"],
+            ["fast-memoize", "npm:2.5.2"],
+            ["wif", "npm:2.0.6"]
+          ],
+          "linkType": "HARD",
+        }]
+      ]],
       ["@arkecosystem/ledger-transport", [
         ["npm:1.1.3", {
           "packageLocation": "./.yarn/cache/@arkecosystem-ledger-transport-npm-1.1.3-3cd458ed90-2.zip/node_modules/@arkecosystem/ledger-transport/",
@@ -516,6 +529,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@arkecosystem/platform-sdk-ark", "workspace:packages/platform-sdk-ark"],
             ["@arkecosystem/core-magistrate-crypto", "npm:2.7.1"],
             ["@arkecosystem/crypto", "npm:2.7.1"],
+            ["@arkecosystem/crypto-identities", "npm:1.1.0"],
             ["@arkecosystem/ledger-transport", "npm:1.1.3"],
             ["@arkecosystem/multi-signature", "npm:0.1.5"],
             ["@arkecosystem/platform-sdk", "workspace:packages/platform-sdk"],
@@ -1254,8 +1268,16 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@arkecosystem/platform-sdk-support", "workspace:packages/platform-sdk-support"],
             ["@ledgerhq/hw-transport-mocker", "npm:5.22.0"],
             ["@liskhq/lisk-client", "npm:2.3.1"],
+            ["@liskhq/lisk-client-new", [
+              "@liskhq/lisk-client",
+              "npm:3.0.0"
+            ]],
             ["@liskhq/lisk-cryptography", "npm:2.4.2"],
             ["@liskhq/lisk-transactions", "npm:2.3.1"],
+            ["@liskhq/lisk-transactions-new", [
+              "@liskhq/lisk-transactions",
+              "npm:3.0.0"
+            ]],
             ["@sindresorhus/tsconfig", "npm:0.7.0"],
             ["@types/eslint", "npm:7.2.2"],
             ["@types/eslint-plugin-prettier", "npm:3.1.0"],
@@ -3475,6 +3497,15 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["axios", "npm:0.19.0"]
           ],
           "linkType": "HARD",
+        }],
+        ["npm:3.0.0", {
+          "packageLocation": "./.yarn/cache/@liskhq-lisk-api-client-npm-3.0.0-f3d2bc080c-2.zip/node_modules/@liskhq/lisk-api-client/",
+          "packageDependencies": [
+            ["@liskhq/lisk-api-client", "npm:3.0.0"],
+            ["@types/node", "npm:12.12.11"],
+            ["axios", "npm:0.19.0"]
+          ],
+          "linkType": "HARD",
         }]
       ]],
       ["@liskhq/lisk-client", [
@@ -3490,6 +3521,19 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@types/node", "npm:10.12.21"]
           ],
           "linkType": "HARD",
+        }],
+        ["npm:3.0.0", {
+          "packageLocation": "./.yarn/cache/@liskhq-lisk-client-npm-3.0.0-c9cf52f697-2.zip/node_modules/@liskhq/lisk-client/",
+          "packageDependencies": [
+            ["@liskhq/lisk-client", "npm:3.0.0"],
+            ["@liskhq/lisk-api-client", "npm:3.0.0"],
+            ["@liskhq/lisk-constants", "npm:1.3.0"],
+            ["@liskhq/lisk-cryptography", "npm:2.4.0"],
+            ["@liskhq/lisk-passphrase", "npm:3.0.0"],
+            ["@liskhq/lisk-transactions", "npm:3.0.0"],
+            ["@types/node", "npm:12.12.11"]
+          ],
+          "linkType": "HARD",
         }]
       ]],
       ["@liskhq/lisk-constants", [
@@ -3498,6 +3542,14 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "packageDependencies": [
             ["@liskhq/lisk-constants", "npm:1.2.3"],
             ["@types/node", "npm:10.12.21"]
+          ],
+          "linkType": "HARD",
+        }],
+        ["npm:1.3.0", {
+          "packageLocation": "./.yarn/cache/@liskhq-lisk-constants-npm-1.3.0-fbc8198825-2.zip/node_modules/@liskhq/lisk-constants/",
+          "packageDependencies": [
+            ["@liskhq/lisk-constants", "npm:1.3.0"],
+            ["@types/node", "npm:12.12.11"]
           ],
           "linkType": "HARD",
         }]
@@ -3510,6 +3562,21 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@liskhq/bignum", "npm:1.3.1"],
             ["@types/ed2curve", "npm:0.2.2"],
             ["@types/node", "npm:10.12.21"],
+            ["buffer-reverse", "npm:1.0.1"],
+            ["ed2curve", "npm:0.2.1"],
+            ["sodium-native", "npm:2.4.6"],
+            ["tweetnacl", "npm:1.0.1"],
+            ["varuint-bitcoin", "npm:1.1.0"]
+          ],
+          "linkType": "HARD",
+        }],
+        ["npm:2.4.0", {
+          "packageLocation": "./.yarn/cache/@liskhq-lisk-cryptography-npm-2.4.0-0bae9efa7a-2.zip/node_modules/@liskhq/lisk-cryptography/",
+          "packageDependencies": [
+            ["@liskhq/lisk-cryptography", "npm:2.4.0"],
+            ["@liskhq/bignum", "npm:1.3.1"],
+            ["@types/ed2curve", "npm:0.2.2"],
+            ["@types/node", "npm:12.12.11"],
             ["buffer-reverse", "npm:1.0.1"],
             ["ed2curve", "npm:0.2.1"],
             ["sodium-native", "npm:2.4.6"],
@@ -3544,6 +3611,16 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["bip39", "npm:2.5.0"]
           ],
           "linkType": "HARD",
+        }],
+        ["npm:3.0.0", {
+          "packageLocation": "./.yarn/cache/@liskhq-lisk-passphrase-npm-3.0.0-6cf9cd18eb-2.zip/node_modules/@liskhq/lisk-passphrase/",
+          "packageDependencies": [
+            ["@liskhq/lisk-passphrase", "npm:3.0.0"],
+            ["@types/bip39", "npm:2.4.1"],
+            ["@types/node", "npm:12.12.11"],
+            ["bip39", "npm:2.5.0"]
+          ],
+          "linkType": "HARD",
         }]
       ]],
       ["@liskhq/lisk-transactions", [
@@ -3556,6 +3633,35 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@types/node", "npm:10.12.21"],
             ["ajv", "npm:6.8.1"],
             ["ajv-merge-patch", "virtual:66cfe6bd64c44b6147a4a8cb1b40c44d578a1c518423400c51726b2b755f24aee6b29fe4e69a612729e1a96f43c557cf378653ac91a25b6ece30d0854925bbc4#npm:4.1.0"]
+          ],
+          "linkType": "HARD",
+        }],
+        ["npm:3.0.0", {
+          "packageLocation": "./.yarn/cache/@liskhq-lisk-transactions-npm-3.0.0-2b615fce83-2.zip/node_modules/@liskhq/lisk-transactions/",
+          "packageDependencies": [
+            ["@liskhq/lisk-transactions", "npm:3.0.0"],
+            ["@liskhq/bignum", "npm:1.3.1"],
+            ["@liskhq/lisk-cryptography", "npm:2.4.0"],
+            ["@liskhq/lisk-validator", "npm:0.3.0"],
+            ["ajv", "npm:6.8.1"],
+            ["ajv-merge-patch", "virtual:66cfe6bd64c44b6147a4a8cb1b40c44d578a1c518423400c51726b2b755f24aee6b29fe4e69a612729e1a96f43c557cf378653ac91a25b6ece30d0854925bbc4#npm:4.1.0"]
+          ],
+          "linkType": "HARD",
+        }]
+      ]],
+      ["@liskhq/lisk-validator", [
+        ["npm:0.3.0", {
+          "packageLocation": "./.yarn/cache/@liskhq-lisk-validator-npm-0.3.0-b37a6a569f-2.zip/node_modules/@liskhq/lisk-validator/",
+          "packageDependencies": [
+            ["@liskhq/lisk-validator", "npm:0.3.0"],
+            ["@liskhq/bignum", "npm:1.3.1"],
+            ["@liskhq/lisk-cryptography", "npm:2.4.0"],
+            ["@types/node", "npm:12.12.11"],
+            ["@types/semver", "npm:5.5.0"],
+            ["@types/validator", "npm:10.9.0"],
+            ["ajv", "npm:6.9.2"],
+            ["semver", "npm:5.6.0"],
+            ["validator", "npm:10.11.0"]
           ],
           "linkType": "HARD",
         }]
@@ -4472,6 +4578,13 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         }]
       ]],
       ["@types/semver", [
+        ["npm:5.5.0", {
+          "packageLocation": "./.yarn/cache/@types-semver-npm-5.5.0-8c205b186f-2.zip/node_modules/@types/semver/",
+          "packageDependencies": [
+            ["@types/semver", "npm:5.5.0"]
+          ],
+          "linkType": "HARD",
+        }],
         ["npm:7.3.3", {
           "packageLocation": "./.yarn/cache/@types-semver-npm-7.3.3-67a3d9ddc0-2.zip/node_modules/@types/semver/",
           "packageDependencies": [
@@ -4513,6 +4626,15 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "packageLocation": "./.yarn/cache/@types-uuid-npm-8.3.0-60b441a3d7-2.zip/node_modules/@types/uuid/",
           "packageDependencies": [
             ["@types/uuid", "npm:8.3.0"]
+          ],
+          "linkType": "HARD",
+        }]
+      ]],
+      ["@types/validator", [
+        ["npm:10.9.0", {
+          "packageLocation": "./.yarn/cache/@types-validator-npm-10.9.0-dce6187fac-2.zip/node_modules/@types/validator/",
+          "packageDependencies": [
+            ["@types/validator", "npm:10.9.0"]
           ],
           "linkType": "HARD",
         }]
@@ -5707,6 +5829,17 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "packageLocation": "./.yarn/cache/ajv-npm-6.8.1-ef7cee5d32-2.zip/node_modules/ajv/",
           "packageDependencies": [
             ["ajv", "npm:6.8.1"],
+            ["fast-deep-equal", "npm:2.0.1"],
+            ["fast-json-stable-stringify", "npm:2.1.0"],
+            ["json-schema-traverse", "npm:0.4.1"],
+            ["uri-js", "npm:4.2.2"]
+          ],
+          "linkType": "HARD",
+        }],
+        ["npm:6.9.2", {
+          "packageLocation": "./.yarn/cache/ajv-npm-6.9.2-ad3edd540c-2.zip/node_modules/ajv/",
+          "packageDependencies": [
+            ["ajv", "npm:6.9.2"],
             ["fast-deep-equal", "npm:2.0.1"],
             ["fast-json-stable-stringify", "npm:2.1.0"],
             ["json-schema-traverse", "npm:0.4.1"],
@@ -15941,6 +16074,13 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         }]
       ]],
       ["semver", [
+        ["npm:5.6.0", {
+          "packageLocation": "./.yarn/cache/semver-npm-5.6.0-e2810885c2-2.zip/node_modules/semver/",
+          "packageDependencies": [
+            ["semver", "npm:5.6.0"]
+          ],
+          "linkType": "HARD",
+        }],
         ["npm:5.7.1", {
           "packageLocation": "./.yarn/cache/semver-npm-5.7.1-40bcea106b-2.zip/node_modules/semver/",
           "packageDependencies": [