+++ packages/platform-sdk-ark/src/contracts.ts
@@ -1,0 +1,4 @@
+export interface CryptoConfig {
+	pubKeyHash: number;
+	wif: number;
+}