+++ packages/platform-sdk-ark/src/services/identity/index.ts
@@ -1,5 +1,5 @@
-import { Managers } from "@arkecosystem/crypto";
 import { Coins, Contracts } from "@arkecosystem/platform-sdk";
+import { CryptoConfig } from "../../contracts";
 import { retrieveCryptoConfiguration } from "../helpers";
 
 import { Address } from "./address";
@@ -10,15 +10,18 @@ import { WIF } from "./wif";
 
 export class IdentityService implements Contracts.IdentityService {
 	readonly #config: Coins.Config;
+	readonly #configCrypto: CryptoConfig;
 
-	private constructor(config: Coins.Config) {
+	private constructor(config: Coins.Config, configCrypto: any) {
 		this.#config = config;
+		this.#configCrypto = {
+			pubKeyHash: configCrypto.network.pubKeyHash,
+			wif: configCrypto.network.wif,
+		};
 	}
 
 	public static async construct(config: Coins.Config): Promise<IdentityService> {
-		await retrieveCryptoConfiguration(config);
-
-		return new IdentityService(config);
+		return new IdentityService(config, (await retrieveCryptoConfiguration(config)).crypto);
 	}
 
 	public async destruct(): Promise<void> {
@@ -26,22 +29,22 @@ export class IdentityService implements Contracts.IdentityService {
 	}
 
 	public address(): Address {
-		return new Address(this.#config);
+		return new Address(this.#config, this.#configCrypto);
 	}
 
 	public publicKey(): PublicKey {
-		return new PublicKey();
+		return new PublicKey(this.#configCrypto);
 	}
 
 	public privateKey(): PrivateKey {
-		return new PrivateKey();
+		return new PrivateKey(this.#configCrypto);
 	}
 
 	public wif(): WIF {
-		return new WIF();
+		return new WIF(this.#configCrypto);
 	}
 
 	public keys(): Keys {
-		return new Keys();
+		return new Keys(this.#configCrypto);
 	}
 }