+++ packages/platform-sdk-profiles/src/mappers/delegate-mapper.test.ts
@@ -30,7 +30,7 @@ beforeAll(() => {
 		.reply(200, require("../../test/fixtures/client/syncing.json"))
 		.get("/api/wallets/D61mfSggzbvQgTUe6JhYKH2doHaqJ3Dyib")
 		.reply(200, require("../../test/fixtures/client/wallet.json"))
-		.get("/api/delegates?page=1")
+		.get("/api/delegates")
 		.reply(200, require("../../test/fixtures/client/delegates-1.json"))
 		.get("/api/delegates?page=2")
 		.reply(200, require("../../test/fixtures/client/delegates-2.json"))