+++ CHANGELOG.md
@@ -5,6 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.9.387 - 2020-10-01
+
+### Fixed
+
+-   **[ARK]** Use `@arkecosystem/crypto-identities` for stateless identity generation (cbe4aab7, @faustbrian)
+
+### Changed
+
+-   **[LSK]** Enable `betanet` support (f2830986, @faustbrian)
+
+## 0.9.385 - 2020-10-01
+
+### Added
+
+-   **[LSK]** Implement API pagination (e1a16729, @faustbrian)
+
+### Changed
+
+-   **[ARK]** Remove compendia testnet (1ef72ebb, @faustbrian)
+-   **[ARK]** Re-apply ARK configurations before every crypto method call (ee26ad7f, @faustbrian)
+-   **[LSK]** Don't enforce api limit and sort transactions by timestamp (09dba752, @dated)
+-   **[LSK]** Handle delegate format in WalletData (7cb9a923, @dated)
+
+### Fixed
+
+-   **[LSK]** Use last response cursor for delegate aggregation (85850d96, @faustbrian)
+
 ## 0.9.379 - 2020-09-30
 
 ### Added